--- src/interfaces/i_lambda_handler_response.ts
@@ -1,5 +1,0 @@
-/** Interface required by AWS Lambda*/
-export interface LambdaResponse {
-  body: string;
-  statusCode: number;
-}
