+++ src/interfaces/index.ts
@@ -1,4 +1,3 @@
-export * from "../merge_request/i_merge_request_event";
+export * from "./i_merge_request_event";
 export * from "./gitlab_api_types";
-export * from "./i_generic_handler_response";
-export * from "./i_lambda_handler_response";
+export * from "./i_lambda_response";
